Detune and Unison
Had to modify the framework to do the unisons, I'm not writing a lot here because I'm sick of looking at this code and it's taken me so long to get that unison going, was working on the detune in parallel as a break from the unison. Both are done now. Need panning still; started on voice count changes although that's a 'nice to have' so whatever